remove captcha request and use random number
Instead of requesting the captcha image and parsing it with tesseract,
now a random number is requested, the response is just an int32 in the
reponse body. This number is sent in the field "numRnd" in every
request.

Note: For each request, a "pre-request" to retrieve a new number is
needed.

Also:
- remove unused code